+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - Added nullable comments in events `Bitrix24AccountBlockedEvent` and `Bitrix24AccountUnblockedEvent`,
   see [add comment to events](https://github.com/bitrix24/b24phpsdk/issues/79).
 - Add result type `Bitrix24\SDK\Services\CRM\Common\Result\SystemFields\Types\File`
+- Add exception `Bitrix24\SDK\Core\Exceptions\ItemNotFoundException`
 - Developer experience: added example `/examples/local-app-workflows` for demonstrate work
   with [workflows](https://apidocs.bitrix24.com/api-reference/bizproc/index.html).
 - Developer experience: added cli make command `make dev-show-fields-description` for show typehints for methods arguments